use a TextEdit instead of a WebView
so now the background of the dict output is transparent instead of
white-on-plasma-background, which was totally fugly

WebView is unusable for us because we can't set a transparent
background. And I'm not going to wait around for the webkit bug
https://bugs.webkit.org/show_bug.cgi?id=40918 to get fixed...

this works great though, it renders HTML. doesn't catch the link clicks
though, and currently it doesn't clip properly. oh and it doesn't allow
flicking/scrollbar yet. but now we can do this...